remove HomologousDensity object
