+++ tardis/visualization/widgets/custom_abundance.py
@@ -14,10 +14,7 @@ import tardis
 from tardis.util.base import quantity_linspace, is_valid_nuclide_or_elem
 from tardis.io.config_reader import Configuration
 from tardis.model import Radial1DModel
-from tardis.model.density import (
-    calculate_power_law_density,
-    calculate_exponential_density,
-)
+from tardis.io.model.density import calculate_power_law_density, calculate_exponential_density
 from tardis.io.config_validator import validate_dict
 from tardis.io.parsers.csvy import load_csvy
 from tardis.io.model_reader import (
